* Fixed npc say * Fixed mounts premium-related stuff * Fixed mounting in PZ with unmountInPz set to false * Fixed exeta vis and all transformItem-related issues (hopefully, please report if anything works like it shouldn't) * Added SQLite to vc10 project (perhaps, can someone remove additional paths which I've commited 2 revs ago or so? Thanks)
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4245 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ weapons.cpp
@@ -877,7 +877,7 @@ void WeaponDistance::onUsedAmmo(Player* player, Item* item, Tile* destTile) cons
 		return;
 
 	if(ammoAction == AMMOACTION_MOVEBACK && breakChance > 0 && random_range(1, 100) < breakChance)
-		g_game.transformItem(item, item->getID(), std::max(0, item->getItemCount() - 1));
+		g_game.transformItem(item, item->getID(), std::max((int32_t)0, (int32_t)item->getItemCount() - 1));
 	else
 		Weapon::onUsedAmmo(player, item, destTile);
 }
